Fix merge done
