+++ src/Response.hpp
@@ -2,13 +2,17 @@
 # define RESPONSE_HPP
 
 # include "webserv.hpp"
+# include "Request.hpp"
 
 class	Response {
 
 public:
 
 	Response(void);
+	Response(Request const & request);
 	virtual	~Response(void);
+
+	void	getHeaders(void);
 
 	void	createHeaders(void);
 
@@ -18,10 +22,14 @@ public:
 
 private:
 
+	std::string	_version;
+	std::string	_status_code;
+	std::string	_status_text;
+	
 	std::map<std::string, std::string>	_headers;
 
 	std::string	_path;
-	std::string	_page;
+	std::string	_body;
 
 };
